requirements.txt
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
huggingface-hub==0.20.2
numpy==1.25.1
openai==0.27.8
scikit-learn==1.3.0
scipy==1.11.1
tokenizers==0.15.0
torch==2.0.1
tqdm==4.65.0
transformers==4.36.2
